Parsoid: Move the Parsoid class to lib/parsoid.js
This will allow us to reuse it in both Parsoid/JS and Parsoid/PHP sys
modules.

Bug: T230791
